portal/requirements.txt
certifi==2018.1.18
chardet==3.0.4
diff-match-patch==20121119
Django==2.0.3
django-filter==1.1.0
django-import-export==1.0.0
et-xmlfile==1.0.1
html5lib==1.0.1
idna==2.6
jdcal==1.4
json-table-schema==0.2.1
lxml==4.2.1
odfpy==1.3.6
openpyxl==2.5.3
python-dateutil==2.7.2
python-magic==0.4.15
pytz==2018.3
PyYAML==3.12
requests==2.18.4
selenium==3.11.0
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
webencodings==0.5.1
xlrd==1.1.0
xlwt==1.3.0
